% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{cum_variance_proportion}
\alias{cum_variance_proportion}
\title{returns the variance for each principal component in pc_dataframe}
\usage{
cum_variance_proportion(pc_dataframe, pc_number = NULL)
}
\arguments{
\item{pc_dataframe}{the dataframe of principal components}

\item{pc_number}{numeric, the desired number of principal components}
}
\value{
a vector of cumulative variance for each principal component
}
\description{
returns the variance for each principal component in pc_dataframe
}
