% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cluster.R
\name{make_cluster}
\alias{make_cluster}
\title{make a cluster class that has the attribute raw and clustered_dt}
\usage{
make_cluster(data, clusterFUN, ...)
}
\arguments{
\item{data}{data to be clustered}

\item{clusterFUN}{cluster function like stats::kmeans}

\item{...}{arguments to be passed into clusterFUN}
}
\value{
a cluster class with a raw dataframe and a clustered dataframe
}
\description{
make a cluster class that has the attribute raw and clustered_dt
}
\examples{
library(stats)

data = data.frame(1:4, c(3,6,8,2), c(4,8,10,2))
result = make_cluster(data, kmeans, centers = 3, nstart = 5, iter.max = 10)
print(result$raw)
print(result$clustered_dt)
}
