% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{pc_name}
\alias{pc_name}
\title{get the name of the PC given the PC number}
\usage{
pc_name(pc_number)
}
\arguments{
\item{pc_number}{the number for the PC}
}
\description{
get the name of the PC given the PC number
}
