% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_helpers.R
\name{pc_names}
\alias{pc_names}
\title{Generates a vector of principal component names given the number of desired principle components}
\usage{
pc_names(pc_number)
}
\arguments{
\item{pc_number}{numeric, the desired number of principal components}
}
\value{
a vector from PC1 to PCX where X is the number of desired PCs
}
\description{
Generates a vector of principal component names given the number of desired principle components
}
