% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot pca labelled by a specified parameter}
\usage{
plot_pca(
  pc_dataframe,
  params,
  cluster = NULL,
  x_pc = 1,
  y_pc = 2,
  point = T,
  label = F,
  used_cols = NULL,
  used_shapes = NULL
)
}
\arguments{
\item{pc_dataframe}{the pca transformed dataframe}

\item{params}{a dataframe with 1 column indicating the param of interest, with colname being param, rownames should be compatible with those of pc_dataframe}

\item{cluster}{(optional), it is the output of make_cluster}

\item{x_pc, }{y_pc the principle components to be ploted on the x or y axis}

\item{point}{if True (default), geom_point will be used}

\item{label}{if True, geom_text will be used}

\item{used_cols, }{used_shapes colours and shapes of the observations}
}
\value{
a ggplot of labelled pca
}
\description{
plot pca labelled by a specified parameter
}
