% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree.R
\name{scree_plot}
\alias{scree_plot}
\title{scree plot}
\usage{
scree_plot(pc_dataframe, pc_number = NULL, cumulative = T)
}
\arguments{
\item{pc_dataframe}{the dataframe of principal components}

\item{pc_number}{numeric, the desired number of principal components}

\item{cumulative}{whether to plot a normal scree plot or a cumulative scree plot}
}
\value{
a ggplot for scree plot
}
\description{
scree plot
}
