% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_within_var_plot.R
\name{tot_within_var_plot}
\alias{tot_within_var_plot}
\title{plot total within cluster variations}
\usage{
tot_within_var_plot(clusters)
}
\arguments{
\item{clusters}{a list of cluster objects generated by make_cluster}
}
\value{
a ggplot of total within cluster variation
}
\description{
plot total within cluster variations
}
\examples{
library(ggplot2)

data = data.frame(1:4, c(3,6,8,2), c(4,8,10,2))
set.seed(1)
n_clusters = 3
clusters = list()
for (i in 1:n_clusters) {
   clusters[[i]] = make_cluster(data, kmeans, centers = i, nstart = 5, iter.max = 10)}
set.seed(NULL)

tot_within_var_plot(clusters = clusters) +
scale_x_continuous(n.breaks = n_clusters)
}
