% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_labelled_cluster_bars.R
\name{plot_labelled_cluster_bars}
\alias{plot_labelled_cluster_bars}
\title{barplot of clusters that's been labelled}
\usage{
plot_labelled_cluster_bars(
  cluster,
  label,
  used_cols = NULL,
  used_shapes = NULL
)
}
\arguments{
\item{cluster}{result from make_cluster}

\item{label}{a dataframe with 1 column indicating the categorical label of interest, with colname being 'label', rownames should be compatible with those of cluster}

\item{used_cols}{(optional) the colours of the clusters}
}
\value{
a ggplot barplot with labels on the x axis and coloured by the clusters
}
\description{
barplot of clusters that's been labelled
}
